+++ ProGetAutomation/Functions/New-ProGetFeed.ps1
@@ -6,12 +6,32 @@ function New-ProGetFeed
     Creates a new ProGet package feed
 
     .DESCRIPTION
-    The `New-ProGetFeed` function creates a new ProGet feed. Use the `Type` parameter to specify the feed type (valid values are 'VSIX', 'RubyGems', 'Docker', 'ProGet', 'Maven', 'Bower', 'npm', 'Deployment', 'Chocolatey', 'NuGet', 'PowerShell'). The `Session` parameter controls the instance of ProGet to connect to. This function uses ProGet's Native API, so an API key is required. Use `New-ProGetSession` to create a session with your API key.
+    The `New-ProGetFeed` function creates a new ProGet feed. Use the `Type` parameter to specify the feed type. The
+    `Session` parameter controls the instance of ProGet to connect to. This function uses ProGet's Native API, so an API
+    key is required. Use `New-ProGetSession` to create a session with your API key.
+
+    [Valid feed types are listed in the Feed Management API documentation.](https://docs.inedo.com/docs/proget-reference-api-feed-management)
+    In April 2023, valid feed types were:
+
+    * asset
+    * chocolatey
+    * docker
+    * helm
+    * maven
+    * npm
+    * nuget
+    * powershell
+    * pypi
+    * romp
+    * rubygems
+    * universal
+    * vsix
 
     .EXAMPLE
     New-ProGetFeed -Session $ProGetSession -Name 'Apps' -Type 'ProGet'
 
-    Demonstrates how to call `New-ProGetFeed`. In this case, a new Universal package feed named 'Apps' will be created for the specified ProGet Uri
+    Demonstrates how to call `New-ProGetFeed`. In this case, a new Universal package feed named 'Apps' will be created
+    for the specified ProGet Uri
     #>
     [CmdletBinding()]
     param(
@@ -25,7 +45,22 @@ function New-ProGetFeed
         [string] $Name,
 
         # The feed type indicates the type of package feed to create.
-        # Valid feed types are ('VSIX', 'RubyGems', 'Docker', 'ProGet', 'Maven', 'Bower', 'npm', 'Deployment', 'Chocolatey', 'NuGet', 'PowerShell') - check here for a latest list - https://inedo.com/support/documentation/proget/feed-types/universal
+        # [Valid feed types are listed in the Feed Management API documentation.](https://docs.inedo.com/docs/proget-reference-api-feed-management)
+        # In April 2023, valid feed types were:
+        #
+        # * asset
+        # * chocolatey
+        # * docker
+        # * helm
+        # * maven
+        # * npm
+        # * nuget
+        # * powershell
+        # * pypi
+        # * romp
+        # * rubygems
+        # * universal
+        # * vsix
         [Parameter(Mandatory)]
         [Alias('FeedType')]
         [string] $Type
@@ -34,13 +69,7 @@ function New-ProGetFeed
     Set-StrictMode -Version 'Latest'
     Use-CallerPreference -Cmdlet $PSCmdlet -SessionState $ExecutionContext.SessionState
 
-    if( -not $Session.ApiKey )
-    {
-        Write-Error -Message ('We are unable to create new package feed ''{0}/{1}'' because your ProGet session is missing an API key. This function uses ProGet''s Native API, which requires an API key. Use `New-ProGetSession` to create a session object that uses an API key.' -f $Type, $Name)
-        return
-    }
-
-    if( $Type -eq 'ProGet' )
+    if ($Type -eq 'ProGet')
     {
         $msg = 'ProGet renamed its "ProGet" feed type name to "Universal". Please update the value of ' +
                'New-ProGetFeed''s "Type" parameter from "ProGet" to "Universal".'
@@ -48,18 +77,22 @@ function New-ProGetFeed
         $Type = 'Universal'
     }
 
-    $Parameters = @{
-                        'FeedType_Name' = $Type;
-                        'Feed_Name' = $Name;
-                    }
+    $Type = $Type.ToLowerInvariant()
 
-    $feedExists = Test-ProGetFeed -Session $Session -Name $Name -Type $Type
-    if( $feedExists )
+    $feedExists = Test-ProGetFeed -Session $Session -Name $Name
+    if ($feedExists)
     {
-        Write-Error -Message ('Unable to create {0} {1} feed: a feed with that name and type already exists.' -f $Type, $Name) -ErrorAction $ErrorActionPreference
+        $msg = "Unable to create ${Type} feed ""${Name}"" because a feed with that name already exists."
+        Write-Error $msg -ErrorAction $ErrorActionPreference
         return
     }
-    Write-Verbose -Message ('Creating {0} {1} feed in ProGet instance "{2}".' -f $Type, $Name, $Session.Url)
-    $null = Invoke-ProGetNativeApiMethod -Session $Session -Name 'Feeds_CreateFeed' -Parameter $Parameters
 
+    $body = [pscustomobject]@{
+        name = $name;
+        feedType = $Type;
+        active = $true;
+    } | ConvertTo-Json
+
+    Write-Information -Message "[$($Session.Url)]  Creating ${Type} feed ""${Name}""."
+    Invoke-ProGetRestMethod -Session $Session -Path '/api/management/feeds/create' -Method Post -Body $body
 }